Move components to /components folder
