+++ src/index.js
@@ -5,7 +5,7 @@ import { createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { logger } from 'redux-logger';
 import rootReducer from './reducers'
-import App from './containers/App'
+import App from './components/App'
 
 const store = createStore(
     rootReducer,
@@ -23,5 +23,5 @@ render(
 );
 
 if (module.hot) {
-  module.hot.accept('./containers/App', () => { render(App) })
+  module.hot.accept('./components/App', () => { render(App) })
 }
